+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `meraki_appliance_vlan_dhcp` resource
 - Add `meraki_organization_license` resource and data source
 - Add `meraki_network_firmware_upgrades` resource and data source
+- Add `meraki_organization_licenses` data source
 
 ## 0.1.2